move dependent ParkContainer interfaces to a global file
